requirements.txt
requests>=2.31.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.14.0
pytest>=7.4.0
pdfplumber>=0.11.0
PyMuPDF>=1.23.0